Lazily resolve whether variable is unique
