+++ org.eclipse.jdt.core.javac/src/org/eclipse/jdt/internal/javac/dom/JavacMethodBinding.java
@@ -31,13 +31,11 @@ import org.eclipse.jdt.core.dom.IMethodBinding;
 import org.eclipse.jdt.core.dom.ITypeBinding;
 import org.eclipse.jdt.core.dom.IVariableBinding;
 import org.eclipse.jdt.core.dom.JavacBindingResolver;
-import org.eclipse.jdt.core.dom.LambdaExpression;
 import org.eclipse.jdt.core.dom.MethodDeclaration;
 import org.eclipse.jdt.core.dom.Modifier;
 import org.eclipse.jdt.core.dom.SingleVariableDeclaration;
 import org.eclipse.jdt.core.dom.TypeParameter;
 import org.eclipse.jdt.core.dom.JavacBindingResolver.BindingKeyException;
-import org.eclipse.jdt.internal.codeassist.DOMCompletionUtil;
 import org.eclipse.jdt.internal.core.BinaryMethod;
 import org.eclipse.jdt.internal.core.JavaElement;
 import org.eclipse.jdt.internal.core.Member;
@@ -688,16 +686,8 @@ public abstract class JavacMethodBinding implements IMethodBinding {
 		if (!this.methodSymbol.isLambdaMethod()) {
 			return new IVariableBinding[0];
 		}
-		LambdaExpression lambdaExpression = (LambdaExpression) this.resolver.symbolToDeclaration.get(this.methodSymbol);
-		MethodDeclaration methodDeclaration = (MethodDeclaration) DOMCompletionUtil.findParent(lambdaExpression, new int[] { ASTNode.METHOD_DECLARATION });
 		return this.methodSymbol.capturedLocals.stream() //
-				.map(varSymbol -> {
-					boolean isUnique = true;
-					if (methodDeclaration != null) {
-						isUnique = JavacBindingResolver.calculateIsUnique(methodDeclaration, getName());
-					}
-					return this.resolver.bindings.getVariableBinding(varSymbol, isUnique);
-				}) //
+				.map(varSymbol -> this.resolver.bindings.getVariableBinding(varSymbol))
 				.toArray(IVariableBinding[]::new);
 	}
 
@@ -746,4 +736,5 @@ public abstract class JavacMethodBinding implements IMethodBinding {
 		}
 		return res;
 	}
+
 }
